#6 validatexmls has been moved to coffee-cdi
